Clean Flags Tests
