+++ playwright_tests/E2E/steps/case-steps.ts
@@ -1,0 +1,7 @@
+import config from "../../config"
+
+export async function routeToCasePage(page, caseId: string) {
+    const casePageUrl = config.CaseBaseURL + '/case-details/' + caseId;
+    console.log("Going to case details page:" + casePageUrl);
+    await page.goto(casePageUrl);
+}
